BBC-399: Pull Request inline comments - created common tables - updated mappings - added manual parsing for comments, because of missing HTML URL on EGIT API - cachted 404 exception for deleted comments - merge from LOCAL_ID to REMOTE_ID and appropriate changes

+++ jira-dvcs-connector-api/src/main/java/com/atlassian/jira/plugins/dvcs/activity/RepositoryActivityDao.java
@@ -25,7 +25,7 @@ public interface RepositoryActivityDao
     
     RepositoryPullRequestMapping findRequestById(int localId);
     
-    RepositoryPullRequestMapping findRequestByRemoteId(int repositoryId, int remoteId);
+    RepositoryPullRequestMapping findRequestByRemoteId(int repositoryId, long remoteId);
 
     Set<String> getExistingIssueKeysMapping(Integer pullRequestId);
 }
